+++ pad/editor/src/backend.rs
@@ -2,15 +2,19 @@ use std::{
     any::Any,
     cell::RefCell,
     collections::{BTreeSet, HashMap},
+    hash::{DefaultHasher, Hash, Hasher},
     io::Cursor,
     path::{Path, PathBuf},
-    sync::Mutex,
+    sync::{
+        atomic::{AtomicUsize, Ordering},
+        Mutex,
+    },
 };
 
 use crate::{get_ast_time, START_TIME};
 use js_sys::Date;
 use leptos::*;
-use uiua::{now, GitTarget, Handle, Report, SysBackend, EXAMPLE_TXT, EXAMPLE_UA};
+use uiua::{now, GitTarget, Handle, Report, Span, SysBackend, Uiua, EXAMPLE_TXT, EXAMPLE_UA};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::JsFuture;
 use web_sys::{HtmlAudioElement, Request, RequestInit, RequestMode, Response};
@@ -20,6 +24,8 @@ pub struct WebBackend {
     pub stderr: Mutex<String>,
     pub trace: Mutex<String>,
     streams: Mutex<HashMap<Handle, VirtualStream>>,
+    id: u64,
+    breakpoint: AtomicUsize,
 }
 
 struct VirtualStream {
@@ -60,14 +66,48 @@ pub fn delete_file(path: &PathBuf) {
     FILES.with(|files| files.borrow_mut().remove(path));
 }
 
+thread_local! {
+    static BREAKPOINTS: RefCell<HashMap<u64, (u64, usize)>> = Default::default();
+}
+
 impl Default for WebBackend {
     fn default() -> Self {
+        Self::new(0, "")
+    }
+}
+
+impl WebBackend {
+    pub fn new(id: impl Hash, code: impl Hash) -> Self {
+        let mut hasher = DefaultHasher::new();
+        id.hash(&mut hasher);
+        let id = hasher.finish();
+        let mut hasher = DefaultHasher::new();
+        code.hash(&mut hasher);
+        let code = hasher.finish();
+        BREAKPOINTS.with(|map| {
+            let mut map = map.borrow_mut();
+            let (old_code, bp) = map.entry(id).or_insert((code, 0));
+            if *old_code != code {
+                *old_code = code;
+                *bp = 0;
+            }
+        });
         Self {
             stdout: Vec::new().into(),
             stderr: String::new().into(),
             trace: String::new().into(),
             streams: HashMap::new().into(),
+            id,
+            breakpoint: AtomicUsize::new(0),
         }
+    }
+    pub fn finish(&self) {
+        BREAKPOINTS.with(|map| {
+            let mut map = map.borrow_mut();
+            if let Some((_, bp)) = map.get_mut(&self.id) {
+                *bp = 0;
+            }
+        });
     }
 }
 
@@ -409,6 +449,34 @@ impl SysBackend for WebBackend {
             return Ok(0.0);
         }
         Ok(-Date::new_0().get_timezone_offset() / 60.0)
+    }
+    fn breakpoint(&self, env: &Uiua) -> Result<bool, String> {
+        let breakpoint = self.breakpoint.fetch_add(1, Ordering::Relaxed);
+        logging::log!("Breakpoint reached: {breakpoint}");
+        let reached = BREAKPOINTS.with(|map| {
+            let mut map = map.borrow_mut();
+            if let Some((_, bp)) = map.get_mut(&self.id) {
+                logging::log!("stored breakpoint: {bp}");
+                let reached = breakpoint >= *bp;
+                if reached {
+                    *bp += 1;
+                }
+                reached
+            } else {
+                false
+            }
+        });
+
+        if !reached {
+            return Ok(true);
+        }
+
+        let message = match env.span() {
+            Span::Code(span) => format!("Break at {span}"),
+            Span::Builtin => "Breakpoint".into(),
+        };
+        self.trace.lock().unwrap().push_str(&message);
+        Ok(false)
     }
 }